+++ HPL2/core/include/impl/GamepadSDL.h
@@ -41,10 +41,10 @@ namespace hpl {
 
 		tString GetGamepadName() { return msGamepadName; }
 
-		int GetNumButtons();
-		int GetNumAxes();
-		int GetNumHats();
-		int GetNumBalls();
+		size_t GetNumButtons();
+		size_t GetNumAxes();
+		size_t GetNumHats();
+		size_t GetNumBalls();
 
 		void Update();
